pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "switchboard"
version = "0.1.0"
description = "Switchboard — MCP↔ACP governance control plane with policy, approvals, and signed provenance."
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Switchboard Contributors"}
]
license = "Apache-2.0"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "httpx>=0.27.0",
  "python-dateutil>=2.9.0",
  "cbor2>=5.6.0",
  "cryptography>=43.0.0",
  "sigstore>=2.1.0",
  "python-dotenv>=1.0.1",
  "opentelemetry-api>=1.26.0",
  "opentelemetry-sdk>=1.26.0",
  "opentelemetry-exporter-otlp>=1.26.0",
  "opentelemetry-instrumentation-fastapi>=0.47b0",
  "structlog>=24.2.0",
  "langgraph>=1.0.0",
  "langchain>=0.2.11",
  "typer>=0.12.4",
  "rich>=13.9.1",
  "jinja2>=3.1.4",
  "pyyaml>=6.0.2",
  "aiosqlite>=0.20.0",
  "asyncpg>=0.29.0",
  "tenacity>=8.4.2",
  "orjson>=3.10.7",
  "streamlit>=1.37.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3",
  "pytest-asyncio>=0.23.8",
  "pytest-cov>=5.0.0",
  "pytest-mock>=3.14.0",
  "httpx>=0.27.0",
  "ruff>=0.6.9",
  "black>=24.8.0",
  "mypy>=1.11.2",
  "types-requests>=2.32.0.20241016",
  "types-pyyaml>=6.0.12.20240917",
  "hypothesis>=6.115.0",
  "pytest-httpx>=0.32.0",
  "tox>=4.23.2",
  "pre-commit>=3.7.1",
  "coverage>=7.6.0",
  "mutmut>=3.0.1",
  "coverage-badge>=1.1.0",
  "mkdocs>=1.6.0"
]

aws = [
  "boto3>=1.35.0",
  "botocore>=1.35.0"
]

gcp = [
  "google-cloud-aiplatform>=1.71.0"
]

[tool.black]
line-length = 100
target-version = ["py311"]
include = "\\.pyi?$"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "C4", "SIM", "ARG"]
ignore = ["E203", "E266", "E501", "B008"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true

[[tool.mypy.overrides]]
module = ["yaml", "yaml.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--strict-markers --disable-warnings -q"
testpaths = [
  "tests"
]

[tool.coverage.run]
branch = true
source = ["switchboard"]

[tool.coverage.report]
skip_empty = true
fail_under = 90

[tool.mutmut]
paths_to_mutate = "switchboard"
tests_dir = "tests"
runner = "python -m pytest"
